fixed kimchi, dick, update client

+++ src/commands/hidden/kimchi.js
@@ -1,5 +1,7 @@
 const { Command } = require('discord-akairo')
-const superagent = require('superagent')
+const isImageUrl = require('is-image-url')
+const redditApiImageGetter = require('reddit-api-image-getter')
+const getter = new redditApiImageGetter()
 
 class KimchiCommand extends Command {
   constructor () {
@@ -20,9 +22,9 @@ class KimchiCommand extends Command {
     const ohNo = await this.client.emojis.resolve('541151482599440385')
     const m = await msg.channel.send(`${loading} **Now subscribed to kimchi facts!**`)
 
-    const response = ['Jugeullae?!', '죽을래']
+    const insult = ['Jugeullae?!', '죽을래']
 
-    if (((msg.author.id == 138549812307034112) || (msg.author.id == 101808227385098240)) == false) return m.edit(response[Math.floor(Math.random() * response.length)])
+    if (((msg.author.id == 138549812307034112) || (msg.author.id == 101808227385098240)) == false) return m.edit(response[Math.floor(Math.random() * insult.length)])
 
     var subreddits = [
       'Nekomimi',
@@ -37,29 +39,15 @@ class KimchiCommand extends Command {
 
     var img_sub = subreddits[Math.round(Math.random() * (subreddits.length - 1))]
 
-    try {
-      var image = await superagent.get(`https://api.imgur.com/3/gallery/r/${img_sub}`).set('authorization', 'Client-ID ' + process.env.IMGUR).then(r => r.body)
-    } catch (e) {
-      return m.edit(`${ohNo} Looks like something went wrong.`).then(msg.delete())
-    }
-
-    if (image.status == 403) {
-      return m.edit(`${ohNo} Looks like something went wrong.`).then(msg.delete())
-    }
-
-    var i = Math.floor(Math.random() * image.data.length)
-
-    if (image.data[i].is_album === true) {
-      var imagePhoto = image.data[i].images[0].link
-    } else {
-      var imagePhoto = image.data[i].link
-    }
+    const response = await getter.getHotImagesOfSubReddit(img_sub)
+    const randomResponse = response[Math.floor(Math.random() * response.length)].url
+    if (isImageUrl(randomResponse) !== true) return m.edit(`${ohNo} Something went wrong, try again.`)
 
     const embed = this.client.util.embed()
       .setTitle('Image didn\'t load click here.')
-      .setURL(imagePhoto)
+      .setURL(randomResponse)
       .setColor(process.env.EMBED)
-      .setImage(imagePhoto)
+      .setImage(randomResponse)
       .setFooter('Requested by REDACTED | via REDACTED • REDATED at XX:XX GMT', 'https://just.vulgarity.xyz/CWtyugHIu6oVFuYN.png')
 
     m.edit({ embed }).then(msg.delete())
